Changes:
- Random generation of request id
- Random generation of server id

+++ Assignment-1/lb/loadbalancer.py
@@ -1,5 +1,6 @@
 from fastapi import FastAPI, Request, HTTPException
 from fastapi.responses import JSONResponse
+from random import randint
 import requests
 import subprocess
 import random
@@ -9,9 +10,9 @@ import uuid
 
 app = FastAPI()
 app.c_hash = ConsistentHashing(3, 512, 9)
-app.request_count = 0
+app.max_request_count = 1e6 # 1 million
 app.serverList = {}
-app.servindex = 1
+app.max_servindex = 1024
 
 
 @app.get("/rep")
@@ -42,7 +43,7 @@ async def add_servers(request: Request):
     has_duplicates = len(hostnames) != len(set(hostnames))
 
     while len(hostnames) < n:
-        cname = uuid.uuid4().hex
+        cname = str(uuid.uuid4().hex)[:6]
         if cname in app.serverList:
             pass
         else:
@@ -89,12 +90,12 @@ async def add_servers(request: Request):
             result = subprocess.run(command, shell=True, text=True)
             print(result.returncode)
             if result.returncode == 0:
-                app.serverList[hostname] = app.servindex
-                app.servindex = app.servindex+1
+                app.serverList[hostname] = randint(1, app.max_servindex)
+                # app.servindex = app.servindex+1
                 ip = subprocess.run(
                     ipcommand, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
                 ipaddr = ip.stdout.strip()
-                app.c_hash.add_server(app.serverList[hostname],ipaddr,8080)
+                app.c_hash.add_server(app.serverList[hostname], ipaddr, 8080)
                 print(ipaddr)
         print(app.serverList)
 
@@ -116,13 +117,47 @@ async def delete_servers(request: Request):
     req = await request.json()
     n = req["n"]
     hostnames = req["hostnames"]
+
+    # picking at random if mentioned list is smaller than n
+    if len(hostnames) < n:
+        available_servers = list(app.serverList.keys())
+        pickserv = [item for item in available_servers if item not in hostnames]
+        if len(pickserv) + len(hostnames) < n:
+            return JSONResponse(
+                status_code=400,
+                content={
+                    "message": "failed as requested deletions exceeds the available number of servers",
+                    "status": "failure"
+                }
+            )
+        else:
+            hostnames += random.sample(pickserv, n-len(hostnames))
+
     if len(hostnames) > n:
         response = {
             "message": "<Error> Length of hostname list is more than newly added instances",
             "status": "failure"
         }
         return JSONResponse(status_code=400, content=response)
+
     else:
+
+        # making a preliminary check whether the requested server delete is possible or not
+        invalidhosts = []
+        for hostname in hostnames:
+            if hostname not in app.serverList:
+                invalidhosts.append(hostname)
+
+        if invalidhosts != []:
+            return JSONResponse(
+                status_code=400,
+                content={
+                    "message": f"The following servers were not found: {', '.join(invalidhosts)}",
+                    "status": "failure"
+                }
+            )
+
+        servers_removed = []
         for hostname in hostnames:
             command = f"docker rm -f {hostname}"
             result = subprocess.run(command, shell=True, text=True)
@@ -131,9 +166,31 @@ async def delete_servers(request: Request):
                 indx = app.serverList[hostname]
                 app.c_hash.remove_server(indx)
                 app.serverList.pop(hostname)
+                servers_removed.append(hostname)
             else:
-                return JSONResponse(status_code=400, content="failed to remove server")
-    return "deleted the server"
+                return JSONResponse(
+                    status_code=400,
+                    content={
+                        "message": {
+                            "N": len(list(app.serverList.keys())),
+                            "replicas": list(app.serverList.keys()),
+                            "error": f"failed to remove server: {hostname}"
+                        },
+                        "status": "failure"
+                    }
+                )
+
+    servers = list(app.serverList.keys())
+    return JSONResponse(
+        status_code=200,
+        content={
+            "message": {
+                "N": len(servers),
+                "replicas": servers
+            },
+            "status": "successful"
+        }
+    )
 
 
 @app.get("/{_path:path}")
@@ -149,9 +206,9 @@ def catch_all_path(_path: str, request: Request):
         url = str(request.url)
         print(url)
         parsed_url = urlparse(url)
-        app.request_count += 1
+        # app.request_count += 1
         server_node = app.c_hash.get_nearest_server(
-            app.c_hash.request_hash(app.request_count))
+            app.c_hash.request_hash(randint(1, app.max_request_count)))
         modified_url = parsed_url._replace(
             netloc=f"{server_node.server_ip}:{server_node.server_port}")
         modified_url = urlunparse(modified_url)
@@ -171,3 +228,8 @@ def catch_all_path(_path: str, request: Request):
             return JSONResponse(content="error occured", status_code=400)
 
         # print(request.client)
+
+if __name__ == "__main__":
+    import uvicorn
+    print("this code is running")
+    uvicorn.run("loadbalancer:app", host="0.0.0.0", port=5000, reload=True)
